* device/include/string.h, * device/lib/_strlen.c: return size_t * device/lib/_memchr.c, * device/lib/_strcspn.c: cosmetic changes * device/lib/sincosf.c, * device/lib/cosf.c, * device/lib/sinf.c: use BOOL instead of larger int * src/z80/gen.c, * src/z80/mappings.i: cosmetic changes * src/SDCCsymt.c (pointerTypes, mergeDeclSpec, checkSClass): fixed bug 2817646   (checkFunction): disabled __naked   (dbuf_printTypeChain): more like RFE 2824498, print C-ish   (isRestrict): added * src/SDCCsymt.h: added IS_RESTRICT, mergeDeclSpec, isRestrict * src/SDCC.y: use mergeDeclSpec to fix bug 2817646 * src/SDCCicode.c (checkPtrQualifiers, checkTypes, geniCodeAssign,   geniCodeParms, geniCodeCall): fixed bug 1302048 * src/SDCCerr.c, * src/SDCCerr.h: added warning W_TARGET_LOST_QUALIFIER * support/regression/tests/bug-136564.c, * support/regression/tests/bug-223113.c, * support/regression/tests/bug1839277.c, * support/regression/tests/bug2655200.c: added necessary const keywords * support/regression/tests/strtok.c: removed incorrect const keyword * support/regression/tests/bug2817646.c: new, added
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@5531 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ device/include/string.h
@@ -74,7 +74,7 @@ extern char *strtok (char * /* restrict*/ str, const char * /*restrict*/ delim);
 /* Miscanelleous functions: */
 extern void *memset (void *s, unsigned char c, size_t n); /* c should be int according to standard. */
 /* extern char *strerror(int errnum); */
-extern int strlen (const char *s); /* Return value should be size_t according to standard. */
+extern size_t strlen (const char *s);
 
 #ifdef SDCC_ds390
 extern void __xdata * memcpyx(void __xdata *, void __xdata *, int) __naked;
